fixes for train position tracking

+++ db/migrate/20100521174200_create_trains.rb
@@ -2,7 +2,7 @@ class CreateTrains < ActiveRecord::Migration
   def self.up
     create_table :trains do |t|
       t.timestamp :last_seen_time
-      t.integer   :travelled_distance, :default => 0
+      t.float     :travelled_distance, :default => 0
       t.string    :line_name
     end
   end
